+++ assets/cursos/curso.js
@@ -1,0 +1,66 @@
+let infoCurso = {
+    'programacion': [
+        {
+            id: 1,
+            titulo: 'aprende python',
+            lenguaje: 'python',
+            vistas: 15000,
+            nivel: 'basico'
+        },
+        {
+            id: 2,
+            titulo: 'aprende python intermedio',
+            lenguaje: 'python',
+            vistas: 142435,
+            nivel: 'intermedio'
+        },
+        {
+            id: 3,
+            titulo: 'aprende python avanzado',
+            lenguaje: 'python',
+            vistas: 34567,
+            nivel: 'avanzado'
+        }
+    ],
+    'matematicas': [
+        {
+            id: 1,
+            titulo: 'aprende matematicas',
+            tema: 'calculo',
+            vistas: 12427,
+            nivel: 'basico'
+        },
+        {
+            id: 2,
+            titulo: 'aprende algebra',
+            tema: 'algebra',
+            vistas: 15722,
+            nivel: 'intermedio'
+        }
+    ],
+    'ingles': [
+        {
+            id: 1,
+            titulo: 'aprende ingles',
+            grado: 'principio',
+            vistas: 11427,
+            nivel: 'basico'
+        },
+        {
+            id: 2,
+            titulo: 'aprende ingles intermedio',
+            grado:'intermedio',
+            vistas: 18722,
+            nivel: 'intermedio'
+        },
+        {
+            id: 3,
+            titulo: 'aprende ingles avanzado',
+            grado:'avanzado',
+            vistas: 234567,
+            nivel: 'avanzado'
+        }
+    ]
+}
+
+module.exports.infoCurso = infoCurso;
